plan subjects

+++ src/plans/plans.service.ts
@@ -1,9 +1,9 @@
 import { Repository } from 'typeorm';
-import { Injectable, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { PlanEntity } from './entities/plan.entity';
 import { CreatePlanDto } from './dto/create-plan.dto';
 import { UpdatePlanDto } from './dto/update-plan.dto';
+import { BadRequestException, Injectable, NotFoundException } from '@nestjs/common';
 
 @Injectable()
 export class PlansService {
@@ -26,9 +26,14 @@ export class PlansService {
   findOne(id: number) {
     return this.repository.findOne({
       where: { id },
-      relations: { subjects: { plan: true, cmk: true }, category: true },
-      select: { category: { id: true }, subjects: true },
+      select: { category: { id: true } },
+      relations: { category: true },
     });
+    // return this.repository.findOne({
+    //   where: { id },
+    //   select: { category: { id: true }, subjects: true },
+    //   relations: { subjects: { plan: true, cmk: true }, category: true },
+    // });
   }
 
   async update(id: number, dto: UpdatePlanDto) {
@@ -46,26 +51,20 @@ export class PlansService {
   }
 
   async remove(id: number) {
+    const plan = await this.repository.findOne({
+      where: { id },
+      relations: { subjects: true },
+      select: { subjects: { id: true } },
+    });
+
+    if (!plan) throw new NotFoundException('План не знайдено');
+
+    if (plan.subjects.length) throw new BadRequestException('Не можливо видалити план в якому є дисципліни');
+
     const res = await this.repository.delete(id);
 
-    if (res.affected === 0) {
-      throw new NotFoundException('Не знайдено');
-    }
+    if (res.affected === 0) throw new NotFoundException('План не знайдено');
 
     return id;
   }
 }
-
-// async findAll() {
-// return this.repository.find({
-//   relations: {
-//     category: true,
-//   },
-// });
-// const qb = await this.repository
-// .createQueryBuilder('plan')
-// .select(['plan.id', 'plan.name'])
-// .leftJoin('plan.category', 'category')
-// .getMany();
-// return qb;
-// }
